server/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
SQLModel==0.0.24
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
Pillow==10.4.0
qrcode==7.4.2
uvicorn==0.30.6
